Fix the code review comments
Add null check for the mandatory parameters and handle the null value in the connector itself.
Declare the variable within the try block in BlobsRetriever class.
Close the fileInputStream even though it is auto closable.
Add the connector name in the package name.
Instead of asserting the one element in the test cases, assert the whole response.